dom/workers/test/serviceworkers/test_navigator.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Bug 930348 - test stub Navigator ServiceWorker utilities.</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
<script class="testbody" type="text/javascript">

  function checkEnabled() {
    ok(navigator.serviceWorker, "navigator.serviceWorker should exist when ServiceWorkers are enabled.");
    ok(typeof navigator.serviceWorker.register === "function", "navigator.serviceWorker.register() should be a function.");
    ok(typeof navigator.serviceWorker.unregister === "function", "navigator.serviceWorker.unregister() should be a function.");
    ok(typeof navigator.serviceWorker.getAll === "function", "navigator.serviceWorker.getAll() should be a function.");
    ok(typeof navigator.serviceWorker.whenReady === "function", "navigator.serviceWorker.whenReady() should be a function.");
    ok(navigator.serviceWorker.installing === null, "There should be no installing worker for an uncontrolled document.");
    ok(navigator.serviceWorker.waiting === null, "There should be no waiting worker for an uncontrolled document.");
    ok(navigator.serviceWorker.current === null, "There should be no current worker for an uncontrolled document.");
  }

  SimpleTest.waitForExplicitFinish();

  SpecialPowers.pushPrefEnv({"set": [
    ["dom.serviceWorkers.enabled", true]
  ]}, function() {
    checkEnabled();
    SimpleTest.finish();
  });
</script>
</pre>
</body>
</html>
